Arrange the file files
