+++ src/Components/Authentication/LogIn.js
@@ -1,14 +1,13 @@
-import { useEffect, useState } from 'react';
-import '../Styles/signup.css';
-import { useDispatch, useSelector } from 'react-redux';
+import { useState, useEffect } from 'react';
+import '../../Styles/signup.css';
 import { Link, useNavigate } from 'react-router-dom';
-import { clearError, createUser } from '../Redux/userSlice';
-import appLogo from './Navigation/logo.png';
+import { useDispatch, useSelector } from 'react-redux';
+import { clearError, loginUser } from '../../Redux/userSlice';
+import appLogo from '../Navigation/logo.png';
 
-const SignUp = () => {
+const LogIn = () => {
   const [formData, setFormData] = useState({
     username: '',
-    email: '',
     password: '',
   });
   const { currentUser, error } = useSelector((state) => state.users);
@@ -19,17 +18,17 @@ const SignUp = () => {
   const handleError = () => (dispatch(clearError()));
   const handleChange = (e) => {
     const { id, value } = e.target;
-    setFormData({
-      ...formData,
+    setFormData((prevFormData) => ({
+      ...prevFormData,
       [id]: value,
-    });
+    }));
     setIsUserExist(false);
     handleError();
   };
 
   const handleSubmit = (e) => {
     e.preventDefault();
-    dispatch(createUser({ user: formData }));
+    dispatch(loginUser({ user: formData }));
   };
   useEffect(() => {
     if (currentUser !== null) {
@@ -43,28 +42,20 @@ const SignUp = () => {
   }, [error]);
   return (
     <div className="outer-cont">
-      <div className="signup-cont">
-        <div className="signup-header">
+      <div className="login-cont">
+        <div className="login-header">
           <div className="logo-container">
             <img id="logo" src={appLogo} alt="app logo" />
           </div>
-          <h1>Register</h1>
+          <h1>Log In</h1>
         </div>
-        {isUserExist && <p className="error-message">User Already Exist!</p>}
-        <form className="signup-form" onSubmit={handleSubmit}>
+        {isUserExist && <p className="error-message">Invalid username or password</p>}
+        <form className="login-form" onSubmit={handleSubmit}>
           <input
             type="text"
             id="username"
             placeholder="Username"
             value={formData.username}
-            onChange={handleChange}
-            required
-          />
-          <input
-            type="email"
-            id="email"
-            placeholder="Email"
-            value={formData.email}
             onChange={handleChange}
             required
           />
@@ -76,16 +67,15 @@ const SignUp = () => {
             onChange={handleChange}
             required
           />
-          <button className="signup-btn" type="submit">SIGN UP</button>
+          <button className="login-btn" type="submit">LOGIN</button>
         </form>
         <div className="footer">
-          <span>Alredy have an account?</span>
-          <Link to="/login" onClick={handleError}>LOGIN</Link>
+          <span>Don&apos;t have an account?</span>
+          <Link to="/signup" onClick={handleError}>SIGN UP</Link>
         </div>
       </div>
     </div>
-
   );
 };
 
-export default SignUp;
+export default LogIn;
