+++ src/styles/models.css
@@ -220,9 +220,9 @@ body {
   }
 
   .btn {
-    width: 4rem;
-    height: 3rem;
-    margin-bottom: 9rem;
+    width: 3.5rem;
+    height: 2.5rem;
+    margin-bottom: 8rem;
   }
 }
 
@@ -295,6 +295,6 @@ body {
   .btn {
     width: 4rem;
     height: 3rem;
-    margin-bottom: 9rem;
+    margin-bottom: 11rem;
   }
 }
